feat: refactor Endpoint components to improve structure and add EndpointSidePanel for better organization
