+++ packages/workbench/src/components/endpoints/endpoint-side-panel.tsx
@@ -1,0 +1,40 @@
+import { Sidebar } from '@/components/sidebar/sidebar'
+import { ApiEndpoint } from '@/types/endpoint'
+import { X } from 'lucide-react'
+import { FC } from 'react'
+import { EndpointBadge } from './endpoint-badge'
+import { EndpointCall } from './endpoint-call'
+import { EndpointDescription } from './endpoint-description'
+
+type Props = { endpoint: ApiEndpoint; onClose: () => void }
+
+export const EndpointSidePanel: FC<Props> = ({ endpoint, onClose }) => {
+  return (
+    <Sidebar
+      initialWidth={600}
+      title={
+        <div className="flex flex-row gap-2 items-center">
+          <EndpointBadge variant={endpoint.method as never}>{endpoint.method.toUpperCase()}</EndpointBadge>
+          <span className="text-md font-bold">{endpoint.path}</span>
+        </div>
+      }
+      onClose={onClose}
+      tabs={[
+        {
+          label: 'Description',
+          content: <EndpointDescription endpoint={endpoint} />,
+        },
+        {
+          label: 'Call',
+          content: <EndpointCall endpoint={endpoint} />,
+        },
+      ]}
+      actions={[
+        {
+          icon: <X className="cursor-pointer w-4 h-4" onClick={onClose} />,
+          onClick: onClose,
+        },
+      ]}
+    />
+  )
+}
